requirements.txt
Flask
gunicorn
twx.botapi==3.1.1
pokebase
fuzzywuzzy